refactor: features 폴더 구조 정리 및 constants 최적화
- domains 폴더를 features로 이름 변경
- 각 feature 내부에 services, constants 폴더로 그룹핑
- constants 파일들을 도메인별로 이관 및 한글 주석으로 변경
- productUtils를 features/product/utils로 이동
- 배럴파일을 통한 import 경로 최적화
- 모든 import 경로를 새로운 구조에 맞게 업데이트
