--- src/basic/domains/product/index.js
@@ -1,2 +1,0 @@
-export * from "./constants";
-export * from "./data";
