+++ src/basic/features/cart/constants/stock.js
@@ -1,0 +1,9 @@
+/**
+ * 재고 관리 관련 상수들
+ */
+
+// 재고 상태 임계값들
+export const STOCK_THRESHOLDS = {
+	LOW_STOCK: 5, // 재고 부족 경고 기준 (5개 미만)
+	CRITICAL_STOCK: 50 // 전체 재고 경고 기준
+};
