+++ src/basic/features/help/components/HelpModalToggleButton.js
@@ -1,0 +1,12 @@
+import { InfoIcon } from "../../../shared/components";
+
+export function HelpModalToggleButton() {
+	return /* HTML */ `
+		<button
+			id="help-modal-toggle"
+			class="fixed top-4 right-4 bg-black text-white p-3 rounded-full hover:bg-gray-900 transition-colors z-50"
+		>
+			${InfoIcon()}
+		</button>
+	`;
+}
