+++ src/basic/features/help/index.js
@@ -1,0 +1,2 @@
+// Help feature exports
+export * from "./components";
