+++ src/basic/features/loyalty/constants/index.js
@@ -1,0 +1,1 @@
+export * from "./points";
