+++ src/basic/features/loyalty/index.js
@@ -1,0 +1,3 @@
+// Loyalty feature exports
+export * from "./services";
+export * from "./constants";
