+++ src/basic/features/loyalty/services/index.js
@@ -1,0 +1,1 @@
+export * from "./loyaltyService";
