+++ src/basic/features/product/components/ProductSelectionPanel.js
