+++ src/basic/features/product/components/StockWarningMessage.js
