+++ src/basic/features/product/constants/index.js
@@ -1,0 +1,1 @@
+export * from "./constants";
