+++ src/basic/features/product/index.js
@@ -1,0 +1,6 @@
+// Product feature exports
+export * from "./components";
+export * from "./constants";
+export * from "./data";
+export * from "./services";
+export * from "./utils";
