+++ src/basic/features/product/services/index.js
@@ -1,0 +1,1 @@
+export * from "./productService";
