+++ src/basic/shared/components/index.js
@@ -1,2 +1,3 @@
+// Re-export all components by category
 export * from "./icons";
 export * from "./layouts";
