+++ src/basic/shared/constants/days.js
@@ -1,0 +1,14 @@
+/**
+ * 요일 관련 상수들
+ */
+
+// 요일별 숫자 매핑 (Date.getDay() 반환값과 매칭)
+export const WEEKDAYS = {
+	SUNDAY: 0,
+	MONDAY: 1, 
+	TUESDAY: 2,
+	WEDNESDAY: 3,
+	THURSDAY: 4,
+	FRIDAY: 5,
+	SATURDAY: 6
+};
